record and reconcile transactions will be from multiple feeding accounts

+++ src_appaserver/delete_folder_row.c
@@ -128,10 +128,37 @@ void delete_where_data_carrot_list_string(
 					  '^' /* delimiter */ ),
 			sql_executable );
 
+	if ( !delete_database )
+	{
+		fprintf( stderr,
+			 "ERROR in %s/%s()/%d: empty delete_database.\n",
+			 __FILE__,
+			 __FUNCTION__,
+			 __LINE__ );
+		exit( 1 );
+	}
+
+	if ( !delete_database->folder )
+	{
+		fprintf( stderr,
+			 "ERROR in %s/%s()/%d: empty folder.\n",
+			 __FILE__,
+			 __FUNCTION__,
+			 __LINE__ );
+		exit( 1 );
+	}
+
 	delete_database->delete_folder_list =
 		delete_database_get_delete_folder_list(
 			application_name,
-			delete_database->folder,
+			delete_database->folder->folder_name,
+			delete_database->folder->post_change_process,
+			delete_database->folder->primary_attribute_name_list,
+			delete_database->folder->row_level_non_owner_forbid,
+			delete_database->folder->row_level_non_owner_view_only,
+			delete_database->folder->mto1_isa_related_folder_list,
+			delete_database->folder->
+				one2m_recursive_related_folder_list,
 			delete_database->dont_delete_mto1_isa,
 			delete_database->primary_attribute_data_list,
 			delete_database->login_name );
